+++ src/libqlocale.cpp
@@ -708,7 +708,7 @@ libqt_string QLocale_CurrencySymbol(const QLocale* self) {
 }
 
 libqt_string QLocale_ToCurrencyString(const QLocale* self, long long param1) {
-    QString _ret = self->toCurrencyString(static_cast<long long>(param1));
+    QString _ret = self->toCurrencyString(static_cast<qlonglong>(param1));
     // Convert QString from UTF-16 in C++ RAII memory to UTF-8 in manually-managed C memory
     QByteArray _b = _ret.toUtf8();
     libqt_string _str;
@@ -720,7 +720,7 @@ libqt_string QLocale_ToCurrencyString(const QLocale* self, long long param1) {
 }
 
 libqt_string QLocale_ToCurrencyString2(const QLocale* self, unsigned long long param1) {
-    QString _ret = self->toCurrencyString(static_cast<unsigned long long>(param1));
+    QString _ret = self->toCurrencyString(static_cast<qulonglong>(param1));
     // Convert QString from UTF-16 in C++ RAII memory to UTF-8 in manually-managed C memory
     QByteArray _b = _ret.toUtf8();
     libqt_string _str;
@@ -1325,7 +1325,7 @@ libqt_string QLocale_CurrencySymbol1(const QLocale* self, int param1) {
 
 libqt_string QLocale_ToCurrencyString22(const QLocale* self, long long param1, const libqt_string symbol) {
     QString symbol_QString = QString::fromUtf8(symbol.data, symbol.len);
-    QString _ret = self->toCurrencyString(static_cast<long long>(param1), symbol_QString);
+    QString _ret = self->toCurrencyString(static_cast<qlonglong>(param1), symbol_QString);
     // Convert QString from UTF-16 in C++ RAII memory to UTF-8 in manually-managed C memory
     QByteArray _b = _ret.toUtf8();
     libqt_string _str;
@@ -1338,7 +1338,7 @@ libqt_string QLocale_ToCurrencyString22(const QLocale* self, long long param1, c
 
 libqt_string QLocale_ToCurrencyString23(const QLocale* self, unsigned long long param1, const libqt_string symbol) {
     QString symbol_QString = QString::fromUtf8(symbol.data, symbol.len);
-    QString _ret = self->toCurrencyString(static_cast<unsigned long long>(param1), symbol_QString);
+    QString _ret = self->toCurrencyString(static_cast<qulonglong>(param1), symbol_QString);
     // Convert QString from UTF-16 in C++ RAII memory to UTF-8 in manually-managed C memory
     QByteArray _b = _ret.toUtf8();
     libqt_string _str;
